cargo +nightly clippy

+++ src/database.rs
@@ -7,9 +7,7 @@ use rocksdb::{
     Direction, Error as RocksDBError, IteratorMode, Options as RocksDBOptions, ReadOptions, DB,
 };
 
-use crate::codec::{encode_data_key, has_prefix, KeyType};
-
-use super::codec::*;
+use crate::codec::*;
 
 /// Database instance.
 pub struct Database {
@@ -50,16 +48,16 @@ pub type Result<T> = core::result::Result<T, Error>;
 
 #[derive(Debug, Clone)]
 pub enum Error {
-    FromUtf8Error(FromUtf8Error),
-    RocksDBError(RocksDBError),
+    FromUtf8(FromUtf8Error),
+    RocksDB(RocksDBError),
     Message(String),
 }
 
 impl std::fmt::Display for Error {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         match self {
-            Error::FromUtf8Error(err) => write!(f, "FromUtf8Error: {}", err),
-            Error::RocksDBError(err) => write!(f, "RocksDBError: {}", err),
+            Error::FromUtf8(err) => write!(f, "FromUtf8Error: {}", err),
+            Error::RocksDB(err) => write!(f, "RocksDBError: {}", err),
             Error::Message(err) => write!(f, "Error: {}", err),
         }
     }
@@ -69,13 +67,13 @@ impl std::error::Error for Error {}
 
 impl From<FromUtf8Error> for Error {
     fn from(e: FromUtf8Error) -> Self {
-        Error::FromUtf8Error(e)
+        Error::FromUtf8(e)
     }
 }
 
 impl From<RocksDBError> for Error {
     fn from(e: RocksDBError) -> Self {
-        Error::RocksDBError(e)
+        Error::RocksDB(e)
     }
 }
 
@@ -147,7 +145,7 @@ impl Database {
 
     pub fn get_or_create_meta(&self, key: &str, key_type: KeyType) -> Result<KeyMeta> {
         let m = self.get_meta(key)?;
-        return match m {
+        match m {
             Some(m) => Ok(m),
             None => {
                 let m = KeyMeta::new(self.next_key_id.get(), key_type);
@@ -155,7 +153,7 @@ impl Database {
                 self.save_meta(key, &m, false)?;
                 Ok(m)
             }
-        };
+        }
     }
 
     pub fn for_each_key<F>(&self, mut f: F) -> Result<usize>
@@ -165,7 +163,7 @@ impl Database {
         let mut counter: usize = 0;
         let mut has_error = None;
         self.prefix_iterator(PREFIX_META, |k, v| {
-            counter = counter + 1;
+            counter += 1;
             match decode_meta_key(k.as_ref()) {
                 Ok(key) => f(key.as_str(), &KeyMeta::from_bytes(v.as_ref())),
                 Err(err) => {
@@ -187,7 +185,7 @@ impl Database {
         let mut counter: usize = 0;
         let mut has_error = None;
         self.prefix_iterator(PREFIX_META, |k, v| {
-            counter = counter + 1;
+            counter += 1;
             if counter > limit {
                 false
             } else {
@@ -220,7 +218,7 @@ impl Database {
             buf
         };
         self.prefix_iterator(k.as_ref(), |k, v| {
-            counter = counter + 1;
+            counter += 1;
             match decode_meta_key(k.as_ref()) {
                 Ok(key) => f(key.as_str(), &KeyMeta::from_bytes(v.as_ref())),
                 Err(err) => {
@@ -561,7 +559,7 @@ impl Database {
     where
         F: FnMut(Box<[u8]>) -> bool,
     {
-        self.for_each_data(key, None, |_, v| f(Box::from(v)))
+        self.for_each_data(key, None, |_, v| f(v))
     }
 
     pub fn list_items(&self, key: &str) -> Result<Vec<Box<[u8]>>> {
@@ -593,26 +591,25 @@ impl Database {
         &self,
         key: &str,
         max_score: Option<&[u8]>,
-    ) -> Result<Option<(Box<[u8]>, Box<[u8]>)>> {
+    ) -> Result<Option<ScoreVal>> {
         let meta = self.get_meta(key)?;
-        let mut ret: Option<(Box<[u8]>, Box<[u8]>)> = None;
         if let Some(mut meta) = meta {
             let (sequence, left_deleted_count, right_deleted_count) =
                 meta.decode_sorted_list_extra();
             let prefix = encode_data_key(meta.id);
             let mut opts = ReadOptions::default();
             opts.set_prefix_same_as_start(true);
-            let iter = self
+            let mut iter = self
                 .rocksdb
                 .iterator_opt(IteratorMode::From(&prefix, Direction::Forward), opts);
-            for (k, v) in iter {
+            if let Some((k, v)) = iter.next() {
                 if !has_prefix(&prefix, k.as_ref()) {
-                    break;
+                    return Ok(None);
                 }
                 let score = decode_data_key_sorted_list_item(k.as_ref());
                 if let Some(max_score) = max_score {
                     if compare_score_bytes(score, max_score) > 0 {
-                        break;
+                        return Ok(None);
                     }
                 }
                 self.rocksdb.delete(k.as_ref())?;
@@ -631,37 +628,35 @@ impl Database {
                     );
                 }
                 self.save_meta(key, &meta, true)?;
-                ret = Some((Box::from(score), v));
-                break;
+                return Ok(Some((Box::from(score), v)));
             }
         }
-        Ok(ret)
+        Ok(None)
     }
 
     pub fn sorted_list_right_pop(
         &self,
         key: &str,
         min_score: Option<&[u8]>,
-    ) -> Result<Option<(Box<[u8]>, Box<[u8]>)>> {
+    ) -> Result<Option<ScoreVal>> {
         let meta = self.get_meta(key)?;
-        let mut ret: Option<(Box<[u8]>, Box<[u8]>)> = None;
         if let Some(mut meta) = meta {
             let (sequence, left_deleted_count, right_deleted_count) =
                 meta.decode_sorted_list_extra();
             let prefix = encode_data_key(meta.id);
             let next_prefix = encode_data_key(meta.id + 1);
             let opts = ReadOptions::default();
-            let iter = self
+            let mut iter = self
                 .rocksdb
                 .iterator_opt(IteratorMode::From(&next_prefix, Direction::Reverse), opts);
-            for (k, v) in iter {
+            if let Some((k, v)) = iter.next() {
                 if !has_prefix(&prefix, k.as_ref()) {
-                    break;
+                    return Ok(None);
                 }
                 let score = decode_data_key_sorted_list_item(k.as_ref());
                 if let Some(min_score) = min_score {
                     if compare_score_bytes(score, min_score) < 0 {
-                        break;
+                        return Ok(None);
                     }
                 }
                 self.rocksdb.delete(k.as_ref())?;
@@ -680,11 +675,10 @@ impl Database {
                     );
                 }
                 self.save_meta(key, &meta, true)?;
-                ret = Some((Box::from(score), v));
-                break;
+                return Ok(Some((Box::from(score), v)));
             }
         }
-        Ok(ret)
+        Ok(None)
     }
 
     pub fn sorted_list_for_each<F>(&self, key: &str, mut f: F) -> Result<u64>
@@ -693,11 +687,11 @@ impl Database {
     {
         self.for_each_data(key, None, |k, v| {
             let score = decode_data_key_sorted_list_item(k.as_ref());
-            f((Box::from(score), Box::from(v)))
+            f((Box::from(score), v))
         })
     }
 
-    pub fn sorted_list_items(&self, key: &str) -> Result<Vec<(Box<[u8]>, Box<[u8]>)>> {
+    pub fn sorted_list_items(&self, key: &str) -> Result<VecScoreVal> {
         let count = self.get_count(key)?;
         let mut vec = Vec::with_capacity(count as u64 as usize);
         self.sorted_list_for_each(key, |item| {
@@ -727,7 +721,7 @@ impl Database {
         })
     }
 
-    pub fn sorted_set_items(&self, key: &str) -> Result<Vec<(Box<[u8]>, Box<[u8]>)>> {
+    pub fn sorted_set_items(&self, key: &str) -> Result<VecScoreVal> {
         let count = self.get_count(key)?;
         let mut vec = Vec::with_capacity(count as u64 as usize);
         self.sorted_set_for_each(key, |v| {
@@ -812,7 +806,7 @@ impl Database {
         key: &str,
         max_score: Option<&[u8]>,
         limit: usize,
-    ) -> Result<Vec<(Box<[u8]>, Box<[u8]>)>> {
+    ) -> Result<VecScoreVal> {
         match self.get_meta(key)? {
             None => Ok(vec![]),
             Some(meta) => {
@@ -850,7 +844,7 @@ impl Database {
         key: &str,
         min_score: Option<&[u8]>,
         limit: usize,
-    ) -> Result<Vec<(Box<[u8]>, Box<[u8]>)>> {
+    ) -> Result<VecScoreVal> {
         match self.get_meta(key)? {
             None => Ok(vec![]),
             Some(meta) => {
